requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
scipy>=1.10.0
sentencepiece>=0.1.99
protobuf>=3.20.0
tensorboard>=2.14.0
tqdm>=4.65.0
numpy>=1.24.0
pyyaml>=6.0
wandb>=0.15.0
bert-score>=0.3.13
matplotlib>=3.5.0
scikit-learn>=1.0.0

